move venti cons to cons.go
