+++ internal/characters/venti/aimed.go
@@ -42,10 +42,7 @@ func (c *char) Aimed(p map[string]int) action.ActionInfo {
 
 	c.Core.QueueAttack(ai, combat.NewCircleHit(c.Core.Combat.Player(), .1, false, combat.TargettableEnemy), aimedHitmark, aimedHitmark+travel)
 	if c.Base.Cons >= 1 {
-		ai.Abil = "Aim (Charged) C1"
-		ai.Mult = ai.Mult / 3.0
-		c.Core.QueueAttack(ai, combat.NewCircleHit(c.Core.Combat.Player(), .1, false, combat.TargettableEnemy), aimedHitmark, aimedHitmark+travel)
-		c.Core.QueueAttack(ai, combat.NewCircleHit(c.Core.Combat.Player(), .1, false, combat.TargettableEnemy), aimedHitmark, aimedHitmark+travel)
+		c.c1(ai, travel)
 	}
 
 	return action.ActionInfo{
