+++ internal/characters/venti/venti.go
@@ -5,11 +5,9 @@ import (
 	"github.com/genshinsim/gcsim/pkg/core"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
 	"github.com/genshinsim/gcsim/pkg/core/combat"
-	"github.com/genshinsim/gcsim/pkg/core/event"
 	"github.com/genshinsim/gcsim/pkg/core/keys"
 	"github.com/genshinsim/gcsim/pkg/core/player/character"
 	"github.com/genshinsim/gcsim/pkg/core/player/character/profile"
-	"github.com/genshinsim/gcsim/pkg/modifier"
 )
 
 func init() {
@@ -45,23 +43,7 @@ func (c *char) Init() error {
 	// C4:
 	// When Venti picks up an Elemental Orb or Particle, he receives a 25% Anemo DMG Bonus for 10s.
 	if c.Base.Cons >= 4 {
-		c.c4bonus = make([]float64, attributes.EndStatType)
-		c.c4bonus[attributes.AnemoP] = 0.25
-		c.Core.Events.Subscribe(event.OnParticleReceived, func(args ...interface{}) bool {
-			// only trigger if Venti catches the particle
-			if c.Core.Player.Active() != c.Index {
-				return false
-			}
-			// apply C4 to Venti
-			c.AddStatMod(character.StatMod{
-				Base:         modifier.NewBaseWithHitlag("venti-c4", 600),
-				AffectedStat: attributes.AnemoP,
-				Amount: func() ([]float64, bool) {
-					return c.c4bonus, true
-				},
-			})
-			return false
-		}, "venti-c4")
+		c.c4()
 	}
 	return nil
 }
